pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neurify-tokenwatch"
version = "0.1.0"
description = "TokenWatch by Neurify — Track, alert, and control your LLM API spending across OpenAI, Anthropic, and Gemini"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Neurify", email = "opensource@neurify.ai"}]
keywords = ["llm", "openai", "anthropic", "gemini", "cost", "budget", "monitoring", "tokenwatch", "neurify"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.24",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
gemini = ["google-generativeai>=0.5"]
all = ["openai>=1.0", "anthropic>=0.20", "google-generativeai>=0.5"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
]

[project.scripts]
tokenwatch = "tokenwatch.cli:main"

[project.urls]
Homepage = "https://github.com/neurify/tokenwatch"
Repository = "https://github.com/neurify/tokenwatch"
Issues = "https://github.com/neurify/tokenwatch/issues"

[tool.hatch.build.targets.wheel]
packages = ["tokenwatch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = false
ignore_missing_imports = true
